% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDEsobol.R
\name{DDEsobol}
\alias{DDEsobol}
\title{Sobol' Sensitivity Analysis for DDE Models}
\usage{
DDEsobol(mod, ...)
}
\arguments{
\item{mod}{either a model function supplied in the manner as needed for 
\code{\link[deSolve]{ode}} (for \code{\link{DDEsobol.default}}) or an 
object of class \code{DDEnetwork} (for \code{\link{DDEsobol.DDEnetwork}}).}

\item{...}{further arguments passed to methods, see \code{\link{DDEsobol.default}} and
\code{\link{DDEsobol.DDEnetwork}}.}
}
\description{
\code{DDEsobol} is the generic function for performing a Sobol' sensitivity
  analysis of DDE models.
}
\details{
There are two methods for this generic function: 
  \code{\link{DDEsobol.default}} (for general DDE models) and
  \code{\link{DDEsobol.DDEnetwork}} (for objects of class \code{DDEnetwork},
  see package \code{DDEnetwork}).
}
\seealso{
\code{\link{DDEsobol.default}, \link{DDEsobol.DDEnetwork}}
}
\author{
Frank Weber
}
