% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEmorris.R
\name{ODEmorris}
\alias{ODEmorris}
\title{Morris SA for objects of class \code{ODEnetwork}}
\usage{
\method{ODEmorris}{ODEnetwork}(odenet, times, ode_method = "lsoda",
  seed = 2015, binf = 0, bsup = 1, r = 25, design = list(type = "oat",
  levels = 100, grid.jump = 1), scale = TRUE)
}
\arguments{
\item{odenet}{[\code{ODEnetwork}]\cr
list of class \code{\link{ODEnetwork}}.}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed
(vector of arbitrary length). Also the
first point of time must be positive.}

\item{ode_method}{[\code{character(1)}]\cr
method to be used for solving the differential equations, see 
\code{\link[deSolve]{ode}}. Defaults to \code{"lsoda"}.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{binf}{[\code{numeric(k)}]\cr
vector of lower borders of possible values for the \code{k} input 
parameters. If they are all equal, a single value can be set.}

\item{bsup}{[\code{numeric(k)}]\cr
vector of upper borders of possible values for the \code{k} input 
parameters. If they are all equal, a single value can be set.}

\item{r}{[\code{integer(1)}]\cr
number of repetitions of the \code{design},
cf. \code{\link[sensitivity]{morris}}.}

\item{design}{[\code{list}]\cr
a list specifying the design type and its parameters,
cf. \code{\link[sensitivity]{morris}}.}

\item{scale}{[\code{logical(1)}]\cr
if \code{TRUE}, scaling is done for the input design of experiments after 
building the design and before calculating the elementary effects,
cf. \code{\link[sensitivity]{morris}}. Defaults to \code{TRUE}, which is
highly recommended if the factors have different orders of magnitude, see
\code{\link[sensitivity]{morris}}.}
}
\value{
List of class \code{morrisRes} with
  \itemize{
    \item \code{res}, a list containing in each element a matrix for one
      output variable. The matrices itself contain the Morris SA results 
      (i.e. \code{mu, mu.star} and \code{sigma} for every parameter) for one 
      point of time per column
    \item \code{pars}, the parameter names.
  }
}
\description{
\code{ODEmorris} performs a sensitivity analysis for objects of class
\code{ODEnetwork} using Morris's elementary effects screening method. 
The sensitivity analysis is done for all output variables, since 
\code{ODEmorris} is an adapted version of \code{\link{ODEmorris_aos}}.
}
\details{
\code{ODEmorris} uses \code{\link[sensitivity]{morris_list}}
which can handle lists as output for its model function. Thus, each
element of the list can be used to contain the results for one output
variable. This saves time since \code{\link[deSolve]{ode}} from the
package \code{deSolve} does its calculations for all output variables anyway,
so \code{\link[deSolve]{ode}} only needs to be executed once.
}
\note{
\code{\link[deSolve]{ode}} sometimes cannot solve an ODE system if 
  unrealistic parameter combinations are sampled by 
  \code{\link[sensitivity]{morris_list}}. Hence \code{NA}s might occur in the 
  Morris sensitivity results, such that \code{\link{ODEmorris_aos}} fails for 
  one or many points of time! For this reason, if \code{NA}s occur, please 
  make use of \code{\link{ODEsobol_ats}} instead or
  restrict the input parameter value intervals usefully using
  \code{binf}, \code{bsup} and \code{scale = TRUE}. It is also helpful to try
  another ODE-solver (argument \code{ode_method}). Problems are known for the
  \code{ode_method}s \code{"euler"}, \code{"rk4"} and \code{"ode45"}. 
  In contrast, the \code{ode_method}s \code{"vode"}, \code{"bdf"}, 
  \code{"bdf_d"}, \code{"adams"}, \code{"impAdams"} and \code{"impAdams_d"} 
  might be even faster than the standard \code{ode_method} \code{"lsoda"}.
  
  If \code{\link[sensitivity]{morris_list}} throws a warning message saying
  "In ... keeping ... repetitions out of ...", try using a bigger number of 
  \code{levels} in the \code{design} argument.
}
\examples{
 # Still missing!

}
\author{
Frank Weber
}
\seealso{
\code{\link[sensitivity]{morris}},
  \code{\link[sensitivity]{morris_list}},
  \code{\link{plot.morrisRes}}
}

