% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODEmorris_ats.R
\name{ODEmorris_ats}
\alias{ODEmorris_ats}
\title{Morris SA for ODEs at All Timepoints Simultaneously}
\usage{
ODEmorris_ats(mod, pars, yini, times, seed = 2015, binf = 0, bsup = 1,
  r = 25, design = list(type = "oat", levels = 100, grid.jump = 1),
  scale = FALSE, ncores = 1)
}
\arguments{
\item{mod}{[\code{function(Time, State, Pars)}]\cr
model to examine, cf. example below.}

\item{pars}{[\code{character(k)}]\cr
vector of \code{k} input variable names.}

\item{yini}{[\code{numeric(z)}]\cr
vector of \code{z} initial values.}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed
(vector of arbitrary length). Also the
first point of time must be positive.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{binf}{[\code{numeric(k)}]\cr
vector of lower borders of possible input parameter values.
If they are all equal, a single value can be set.}

\item{bsup}{[\code{numeric(k)}]\cr
vector of upper borders of possible input parameter values.
If they are all equal, a single value can be set.}

\item{r}{[\code{integer(1)}]\cr
number of repetitions of the \code{design},
cf. \code{\link[sensitivity]{morris}}.}

\item{design}{[\code{list}]\cr
a list specifying the design type and its parameters,
cf. \code{\link[sensitivity]{morris}}.}

\item{ncores}{[\code{integer(1)}]\cr
  number of processor cores to be used for calculating the sensitivity
  indices. Must be between 1 and 4.}
}
\value{
list of class \code{morrisRes_ats} with
  \itemize{
    \item \code{res}, the list of Morris SA results for every output
      variable (each element of the list contains a matrix of
      \code{mu, mu.star} and \code{sigma} for every parameter and
      every point of time in the \code{times} vector)
    \item \code{pars}, the parameter names.
  }
}
\description{
\code{ODEmorris_ats} performs a sensitivity analysis for
ordinary differential equations at all timepoints
simultaneously using Morris's elementary effects
screening method.
}
\note{
\code{\link[deSolve]{ode}} or rather its standard solver \code{lsoda}
  sometimes cannot solve an ODE system if unrealistic parameters
  are sampled by \code{\link[sensitivity]{morris}}. Hence
  \code{NA}s might occur in the Morris sensitivity results, such
  that \code{\link{ODEmorris}} fails for one or many points of time!
  For this reason, if \code{NA}s occur, please make use of
  \code{\link{ODEsobol}} instead or
  restrict the input parameter value intervals usefully using
  \code{binf} and \code{bsup}!
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al, 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # parameter a
              b = 0.3,     # parameter b
              s = 3)       # parameter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 100, by = 10)

FHNres <- ODEmorris(mod = FHNmod,
                    pars = names(FHNpars),
                    yini = FHNyini,
                    times = FHNtimes,
                    seed = 2015,
                    binf = c(0.18, 0.18, 2.8),
                    bsup = c(0.22, 0.22, 3.2),
                    r = 25,
                    design =
                        list(type = "oat", levels = 100, grid.jump = 1),
                    ncores = 4)
}
\seealso{
\code{\link[sensitivity]{morris}},
  \code{\link{plot.morrisRes}}
}

