\name{ODEsensitivity-package}
\alias{ODEsensitivity-package}
\alias{ODEsensitivity}
\docType{package}
\title{
\packageTitle{ODEsensitivity}
}
\description{
\packageDescription{ODEsensitivity}
}
\details{
\packageDESCRIPTION{ODEsensitivity}
\packageIndices{ODEsensitivity}
}
\author{
Stefan Theers
Dirk Surmann

Maintainer: Stefan Theers <stefan.theers@tu-dortmund.de>
}
\references{
Saltelli et al. (2010): Sensitivity Analysis
others.
}
\keyword{ package }
\seealso{
\link[sensitivity]{morris}, \link[sensitivity]{sobol},
\link[deSolve]{ode}
}
\examples{
x <- 2
}
