% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODEsensitivity-package.R
\docType{package}
\name{ODEsensitivity}
\alias{ODEsensitivity}
\alias{ODEsensitivity-package}
\title{Performing Sensitivity Analysis in ODE Models}
\description{
ODEsensitivity provides methods to perform sensitivity analysis (SA)
in ordinary differential equation (ODE) models.
\code{\link{ODEmorris}} and \code{\link{ODEsobol}}
are based on the implementations of Morris and Sobol SA in the
\link{sensitivity} package (Pujol et al., 2015).
}
\details{
cf. \link{sensitivity} package and its \code{\link[sensitivity]{morris}}
or \code{\link[sensitivity]{soboljansen}} implementations.
}

