% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODEsobol.R
\name{ODEsobol}
\alias{ODEsobol}
\title{Sobol SA for ODEs}
\usage{
ODEsobol(mod, pars, yini, times, seed = 2015, n = 1000,
  trafo = function(Y) rowSums(Y^2))
}
\arguments{
\item{mod}{model to examine.}

\item{pars}{vector of input variable names.}

\item{yini}{vector of initial values.}

\item{times}{points of time at which the SA should be executed.}

\item{seed}{seed.}

\item{n}{parameter \code{nboot} used in \code{\link{sobol2007}},
i.e. the number of bootstrap replicates.}

\item{trafo}{function to transform \code{z > 1} output variables to
  IR [only needed, if \code{z > 1}]. Must be able to deal with a
  matrix.}
}
\value{
list of Sobol SA results (i.e. 1st order sensitivity indices
  \code{S} and total sensitivity indices \code{T}) for every point of
  time of the \code{times} vector.
}
\description{
\code{ODEsobol} performs a sensitivity analysis for
ordinary differential equations using
the variance-based Sobol method.
}
\examples{
x <- 5
}
\seealso{
\code{\link[sensitivity]{sobol}},
  \code{\link[sensitivity]{sobol2007}}
}

