% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEsobol_aos.R
\name{ODEsobol_aos}
\alias{ODEsobol_aos}
\title{Sobol SA for ODEs for All Output Variables and All Timepoints
Simultaneously}
\usage{
ODEsobol_aos(mod, pars, yini, times, ode_method = "lsoda", seed = 2015,
  n = 1000, rfuncs = rep("runif", length(pars)),
  rargs = rep("min = 0, max = 1", length(pars)), method = "martinez",
  nboot = 0)
}
\arguments{
\item{mod}{[\code{function(Time, State, Pars)}]\cr
model to examine, cf. example below.}

\item{pars}{[\code{character(k)}]\cr
vector of \code{k} input variable names.}

\item{yini}{[\code{numeric(z)}]\cr
vector of \code{z} initial values.}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed
(vector of arbitrary length). Also the
first point of time must be positive.}

\item{ode_method}{[\code{character(1)}]\cr
method to be used for solving the differential equations, see 
\code{\link[deSolve]{ode}}. Defaults to \code{"lsoda"}.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{n}{[\code{integer(1)}]\cr
number of random parameter values (\code{n} per input factor) used to 
estimate the variance-based sensitivity indices by Monte-Carlo-method.
(Variance-based methods for sensitivity analysis rely on 
Monte-Carlo-simulation to estimate the integrals needed for the calculation
of the sensitivity indices.) Defaults to 1000.}

\item{rfuncs}{[\code{character(k)}]\cr
names of the \code{k} functions used to generate the \code{n} random values
for the \code{k} parameters. This way, different distributions can be 
used for the \code{k} parameters. Defaults to \code{"runif"} for each of
the \code{k} parameters.}

\item{rargs}{[\code{character(k)}]\cr
arguments to be passed to the \code{k} functions of \code{rfuncs}. Each 
element of \code{rargs} has to be a string of type \code{"tag1 = value1, 
tag2 = value2, ..."}. By default, \code{min = 0} and \code{max = 1} are 
used for each of the \code{k} \code{runif}'s, meaning a uniform 
distribution of all parameters on [0, 1].}

\item{method}{[\code{character(1)}]\cr
either \code{"jansen"} or \code{"martinez"}, specifying which modification
of the variance-based Sobol method shall be used. Defaults to 
\code{"martinez"}, which is slightly faster than \code{"jansen"}.}

\item{nboot}{[\code{integer(1)}]\cr
parameter \code{nboot} used in \code{\link{soboljansen_matrix}} resp.
\code{\link{sobolmartinez_matrix}}, i.e. the number of bootstrap 
replicates. Defaults to 0, so no bootstrapping is done.}
}
\value{
List of Sobol SA results (i.e. 1st order sensitivity indices
  \code{S} and total sensitivity indices \code{T}) for every point of
  time of the \code{times} vector, of class \code{sobolRes_ats}.
}
\description{
\code{ODEsobol_aos} performs a variance-based sensitivity analysis for
ordinary differential equations according to the Sobol-Martinez-method. 
The analysis is done for all output variables at all timepoints 
simultaneously using \code{\link[sensitivity]{sobolmartinez_list}} from the 
package \code{sensitivity}.
}
\details{
\code{ODEmorris_aos} uses 
\code{\link[sensitivity]{sobolmartinez_list}} which can handle lists 
as output for its model function. Thus, each element of the list can be used
to contain the results for one output variable. This saves time since 
\code{\link[deSolve]{ode}} from the package \code{deSolve} does its 
calculations for all output variables anyway, so \code{\link[deSolve]{ode}} 
only needs to be executed once.
}
\note{
Sometimes, it is also helpful to try another ODE-solver (argument 
  \code{ode_method}). Problems are known for the
  \code{ode_method}s \code{"euler"}, \code{"rk4"} and \code{"ode45"}. 
  In contrast, the \code{ode_method}s \code{"vode"}, \code{"bdf"}, 
  \code{"bdf_d"}, \code{"adams"}, \code{"impAdams"} and \code{"impAdams_d"} 
  might be even faster than the standard \code{ode_method} \code{"lsoda"}.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 20, by = 0.5)

FHNres <- ODEsobol_aos(mod = FHNmod,
                       pars = c("a", "b", "s"),
                       yini = FHNyini,
                       times = FHNtimes,
                       ode_method = "adams",
                       seed = 2015,
                       n = 10,               # use n >> 10!
                       rfuncs = c("runif", "runif", "rnorm"),
                       rargs = c(rep("min = 0.18, max = 0.22", 2),
                                 "mean = 3, sd = 0.2 / 3"),
                       method = "martinez",
                       nboot = 0)

}
\author{
Frank Weber
}
\seealso{
\code{\link[sensitivity]{sobol}, 
\link[sensitivity]{sobolmartinez_list},
\link{plot.sobolRes_aos}}
}

