% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEsobol_ats.R
\name{ODEsobol_ats}
\alias{ODEsobol_ats}
\title{Sobol' SA for ODEs at All Timepoints Simultaneously}
\usage{
ODEsobol_ats(mod, pars, yini, times, y_analyzed = names(yini)[1],
  ode_method = "lsoda", seed = 2015, n = 1000, rfuncs = rep("runif",
  length(pars)), rargs = rep("min = 0, max = 1", length(pars)),
  method = "martinez", nboot = 0)
}
\arguments{
\item{mod}{[\code{function(Time, State, Pars)}]\cr
model to examine, cf. example below.}

\item{pars}{[\code{character(k)}]\cr
vector of \code{k} input variable names.}

\item{yini}{[\code{numeric(z)}]\cr
vector of \code{z} initial values. Must be named (with unique names).}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed
(vector of arbitrary length). Also the
first point of time must be positive.}

\item{y_analyzed}{[\code{character(1)}]\cr
name of the \code{yini}-variable to be analyzed. Defaults to the name of 
the first \code{yini}-variable.}

\item{ode_method}{[\code{character(1)}]\cr
method to be used for solving the differential equations, see 
\code{\link[deSolve]{ode}}. Defaults to \code{"lsoda"}.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{n}{[\code{integer(1)}]\cr
number of random parameter values (\code{n} per input factor) used to 
estimate the variance-based sensitivity indices by Monte-Carlo-method.
(Variance-based methods for sensitivity analysis rely on 
Monte-Carlo-simulation to estimate the integrals needed for the calculation
of the sensitivity indices.) Defaults to 1000.}

\item{rfuncs}{[\code{character(k)}]\cr
names of the \code{k} functions used to generate the \code{n} random values
for the \code{k} parameters. This way, different distributions can be 
used for the \code{k} parameters. Defaults to \code{"runif"} for each of
the \code{k} parameters.}

\item{rargs}{[\code{character(k)}]\cr
arguments to be passed to the \code{k} functions of \code{rfuncs}. Each 
element of \code{rargs} has to be a string of type \code{"tag1 = value1, 
tag2 = value2, ..."}. By default, \code{min = 0} and \code{max = 1} are 
used for each of the \code{k} \code{runif}'s, meaning a uniform 
distribution of all parameters on [0, 1].}

\item{method}{[\code{character(1)}]\cr
either \code{"jansen"} or \code{"martinez"}, specifying which modification
of the variance-based Sobol' method shall be used. Defaults to 
\code{"martinez"}, which is slightly faster than \code{"jansen"}.}

\item{nboot}{[\code{integer(1)}]\cr
parameter \code{nboot} used in \code{\link{soboljansen_matrix}} resp.
\code{\link{sobolmartinez_matrix}}, i.e. the number of bootstrap 
replicates. Defaults to 0, so no bootstrapping is done.}
}
\value{
List of Sobol' SA results (i.e. 1st order sensitivity indices
  \code{S} and total sensitivity indices \code{T}) for every point of
  time of the \code{times} vector, of class \code{sobolRes_ats}.
}
\description{
\code{ODEsobol_ats} performs a variance-based sensitivity analysis for
ordinary differential equations according to either the Sobol'-Jansen- or the
Sobol'-Martinez-method. The analysis is done for one output variable at all 
timepoints simultaneously using \code{\link[sensitivity]{soboljansen_matrix}}
resp. \code{\link[sensitivity]{sobolmartinez_matrix}} from the package 
\code{sensitivity}.
}
\details{
\code{ODEsobol_ats} is faster than \code{ODEsobol} since 
\code{\link[sensitivity]{soboljansen_matrix}} and
\code{\link[sensitivity]{sobolmartinez_matrix}} can handle matrix output for 
their model functions. In \code{ODEsobol_ats}, a model function is
created returning a matrix with the \code{\link[deSolve]{ode}}-results for 
all timepoints (one per column). Thus, \code{\link[deSolve]{ode}} only needs
to be executed once.
}
\note{
\code{ODEsobol_ats} is only purposed for analyzing one output 
variable.

  Sometimes, it is also helpful to try another ODE-solver (argument 
  \code{ode_method}). Problems are known for the
  \code{ode_method}s \code{"euler"}, \code{"rk4"} and \code{"ode45"}. 
  In contrast, the \code{ode_method}s \code{"vode"}, \code{"bdf"}, 
  \code{"bdf_d"}, \code{"adams"}, \code{"impAdams"} and \code{"impAdams_d"} 
  might be even faster than the standard \code{ode_method} \code{"lsoda"}.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007)
# definition of the model itself, parameters, initial values
# and the times vector:

}
\author{
Frank Weber
}
\references{
J. O. Ramsay, G. Hooker, D. Campbell and J. Cao, 2007,
  \emph{Parameter estimation for differential equations: a generalized 
  smoothing approach}, Journal of the Royal Statistical Society, Series B, 
  69, Part 5, 741--796.
}
\seealso{
\code{\link[sensitivity]{soboljansen_matrix},
\link[sensitivity]{sobolmartinez_matrix},
\link{plot.sobolRes_ats}}
}

