% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morrisRes_aos.R
\name{plot.morrisRes_aos}
\alias{plot.morrisRes_aos}
\title{Plotting the results of Morris SA for objects of class \code{morrisRes_aos}}
\usage{
\method{plot}{morrisRes_aos}(x, y_idx = 1, type = "sep",
  legendPos = "topleft", ...)
}
\arguments{
\item{x}{[\code{morrisRes_aos}]\cr
resulting output of \code{\link{ODEmorris_aos}}, of class 
\code{morrisRes_aos}.}

\item{y_idx}{[\code{integer(1)}]\cr
index of the output variable to be plotted. Defaults to 1.}

\item{type}{[\code{character(1)}]\cr
plot type, choose between \code{"sep"} and \code{"trajec"}.}

\item{legendPos}{[\code{character(1)}]\cr
legend position, default is \code{"topleft"}.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
TRUE (invisible; for testing purposes).
}
\description{
\code{plot} plots the results of Morris SA for objects of class 
\code{morrisRes_aos}.
}
\details{
\code{plot} with \code{type = "sep"} plots mu.star and
  sigma separately versus time.

\code{plot} with \code{type = "trajec"} plots mu.star versus
  sigma for every point of time.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # parameter a
              b = 0.3,     # parameter b
              s = 3)       # parameter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 100, by = 10)

FHNres_aos <- ODEmorris_aos(mod = FHNmod,
                            pars = names(FHNpars),
                            yini = FHNyini,
                            times = FHNtimes,
                            seed = 2015,
                            binf = c(0.18, 0.18, 2.8),
                            bsup = c(0.22, 0.22, 3.2),
                            r = 25,
                            design =
                               list(type = "oat", levels = 100, 
                                    grid.jump = 1),
                            scale = TRUE)

# Plots:
plot(FHNres_aos, y_idx = 1, type = "sep")
plot(FHNres_aos, y_idx = 1, type = "trajec")
plot(FHNres_aos, y_idx = 2, type = "sep")
plot(FHNres_aos, y_idx = 2, type = "trajec")

}
\author{
Stefan Theers
}
\seealso{
\code{\link{ODEmorris_aos}},
  \code{\link[sensitivity]{morris_list}}
}

