% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morrisRes_ats.R
\name{plot.morrisRes_ats}
\alias{plot.morrisRes_ats}
\title{Plotting the results of Morris SA for objects of class \code{morrisRes_ats}}
\usage{
\method{plot}{morrisRes_ats}(x, type = "sep", main_title = NULL,
  legendPos = "topleft", ...)
}
\arguments{
\item{x}{[\code{morrisRes_ats}]\cr
resulting output of \code{\link{ODEmorris_ats}}, of class 
\code{morrisRes_ats}.}

\item{type}{[\code{character(1)}]\cr
plot type, choose between \code{"sep"} and \code{"trajec"}.}

\item{main_title}{[\code{character(1)}]\cr
title for the plot. If \code{type = "sep"}, this is the overall title for
the two separate plots. Defaults to NULL, so a standard title is generated.}

\item{legendPos}{[\code{character(1)}]\cr
legend position, default is \code{"topleft"}.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
TRUE (invisible; for testing purposes).
}
\description{
\code{plot} plots the results of Morris SA for objects of class 
\code{morrisRes_ats}.
}
\details{
\code{plot} with \code{type = "sep"} plots mu.star and
  sigma separately versus time.

\code{plot} with \code{type = "trajec"} plots mu.star versus
  sigma for every point of time.
}
\note{
Unfortunately, the passing of arguments (e.g. "main") does not work
  correctly.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 50, by = 5)

FHNres_ats <- ODEmorris_ats(mod = FHNmod,
                            pars = c("a", "b", "s"),
                            yini = FHNyini,
                            times = FHNtimes,
                            ode_method = "adams",
                            y_idx = 1,        # voltage only
                            seed = 2015,
                            binf = c(0.18, 0.18, 2.8),
                            bsup = c(0.22, 0.22, 3.2),
                            r = 25,
                            design =
                              list(type = "oat", levels = 100, 
                                   grid.jump = 1),
                            scale = TRUE)  

# Plots:
plot(FHNres_ats, type = "sep")
plot(FHNres_ats, type = "trajec")

}
\author{
Frank Weber
}
\seealso{
\code{\link{ODEmorris_ats}},
  \code{\link[sensitivity]{morris_matrix}}
}

