% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sobolRes.R
\name{plot.sobolRes}
\alias{plot.sobolRes}
\title{Plotting the results of Sobol' SA for objects of class \code{sobolRes}}
\usage{
\method{plot}{sobolRes}(x, y_plot = names(x$ST_by_y)[1], type = "b",
  main_title = NULL, legendPos = "topleft", ...)
}
\arguments{
\item{x}{[\code{sobolRes}-object]\cr
resulting output of \code{\link{ODEsobol}}, of class 
\code{sobolRes}.}

\item{y_plot}{[\code{character(1)}]\cr
name of the \code{yini}-variable to be plotted. Defaults to the name of 
the first \code{yini}-variable.}

\item{type}{[\code{character(1)}]\cr
plot type, i.e. \code{"p", "l", "b", "c", "o", "s", "h"} or \code{"n"}. 
Defaults to \code{"b"}.}

\item{main_title}{[\code{character(1)}]\cr
the common title for the two graphics. Default is \code{NULL}, which means
an automatic title is generated.}

\item{legendPos}{[\code{character(1)}]\cr
legend position, default is \code{"topleft"}.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
TRUE (invisible; for testing purposes).
}
\description{
\code{plot.sobolRes} plots the results of Sobol' SA for objects of class 
\code{sobolRes}.
}
\details{
1st order and total Sobol' SA indices are plotted for one 
\code{yini}-variable (chosen by argument \code{y_plot}) and for each input 
parameter against time.
}
\note{
Unfortunately, the passing of arguments (e.g. "main") does not work
  correctly.
}
\examples{
# Missing.

}
\author{
Frank Weber
}
\seealso{
\code{\link{ODEsobol}, \link[sensitivity]{sobolmartinez_list}}
}

