% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sobolRes_ats.R
\name{plot.sobolRes_ats}
\alias{plot.sobolRes_ats}
\title{Plotting the results of Sobol SA}
\usage{
\method{plot}{sobolRes_ats}(x, type = "b", legendPos = "topleft", ...)
}
\arguments{
\item{x}{[\code{sobolRes_ats}]\cr
resulting output of \code{\link{ODEsobol_ats}}, of class 
\code{sobolRes_ats}.}

\item{type}{[\code{character(1)}]\cr
plot type, i.e. \code{"p", "l", "b", "c"} or \code{"n"}.}

\item{legendPos}{[\code{character(1)}]\cr
legend position, default is \code{"topleft"}.}

\item{...}{additional arguments.}
}
\value{
TRUE (invisible; for testing purposes).
}
\description{
\code{plot.sobolRes_ats} plots the results of Sobol SA.
}
\details{
1st order and total Sobol SA indices are plotted for each input
parameter against time.
}
\examples{

##### FitzHugh-Nagumo equations (Ramsay et al., 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # parameter a
              b = 0.3,     # parameter b
              s = 3)       # parameter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 20, by = 0.5)

FHNres_ats <- ODEsobol_ats(mod = FHNmod,
                           pars = names(FHNpars),
                           yini = FHNyini,
                           times = FHNtimes,
                           y_idx = 1,
                           seed = 2015,
                           n = 10)               # use n >> 10!

# Plot:
plot(FHNres_ats, type = "l", legendPos = "topright")

}
\author{
Frank Weber
}
\seealso{
\code{\link{ODEsobol_ats}},
  \code{\link[sensitivity]{sobol}},
  \code{\link[sensitivity]{soboljansen_matrix}}
}

