% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sobolRes_trafo.R
\name{plot.sobolRes_trafo}
\alias{plot.sobolRes_trafo}
\title{Plotting the results of Sobol' SA for objects of class \code{sobolRes_trafo}}
\usage{
\method{plot}{sobolRes_trafo}(x, type = "b", main_title = NULL,
  legendPos = "topleft", ...)
}
\arguments{
\item{x}{[\code{sobolRes}]\cr
resulting output of \code{\link{ODEsobol_trafo}}, of class 
\code{sobolRes_trafo}.}

\item{type}{[\code{character(1)}]\cr
plot type, i.e. \code{"p", "l", "b", "c", "o", "s", "h"} or \code{"n"}. 
Defaults to \code{"b"}.}

\item{main_title}{[\code{character(1)}]\cr
the common title for the two graphics. Default is \code{NULL}, which means
an automatic title is generated.}

\item{legendPos}{[\code{character(1)}]\cr
legend position, default is \code{"topleft"}.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
TRUE (invisible; for testing purposes).
}
\description{
\code{plot.sobolRes_trafo} plots the results of Sobol' SA for objects of 
class \code{sobolRes_trafo}.
}
\details{
1st order and total Sobol' SA indices are plotted for each input
parameter against time.
}
\note{
Unfortunately, the passing of arguments (e.g. "main") does not work
  correctly.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007)
# definition of the model itself, parameters, initial values
# and the times vector:

}
\author{
Stefan Theers
}
\seealso{
\code{\link{ODEsobol_trafo}},
  \code{\link[sensitivity]{soboljansen}},
  \code{\link[sensitivity]{sobolmartinez}}
}

