% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEmorris.default.R
\name{ODEmorris.default}
\alias{ODEmorris.default}
\title{Morris SA for General ODE Models}
\usage{
\method{ODEmorris}{default}(mod, pars, state_init, times, seed = 2015,
  binf = 0, bsup = 1, r = 25, design = list(type = "oat", levels = 100,
  grid.jump = 1), scale = TRUE, ode_method = "lsoda",
  ode_parallel = FALSE, ode_parallel_ncores = NA, ...)
}
\arguments{
\item{mod}{[\code{function(Time, State, Pars)}]\cr
model to examine, cf. example below.}

\item{pars}{[\code{character(k)}]\cr
vector of \code{k} input variable names.}

\item{state_init}{[\code{numeric(z)}]\cr
vector of \code{z} initial values. Must be named (with unique names).}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed (vector of arbitrary 
length). The first point of time must be greater than zero.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{binf}{[\code{numeric(k)}]\cr
vector of lower borders of possible input parameter values.
If they are all equal, a single value can be set.}

\item{bsup}{[\code{numeric(k)}]\cr
vector of upper borders of possible input parameter values.
If they are all equal, a single value can be set.}

\item{r}{[\code{integer(1)}]\cr
number of repetitions of the \code{design},
cf. \code{\link[sensitivity]{morris}}.}

\item{design}{[\code{list}]\cr
a list specifying the design type and its parameters,
cf. \code{\link[sensitivity]{morris}}.}

\item{scale}{[\code{logical(1)}]\cr
if \code{TRUE}, scaling is done for the input design of experiments after 
building the design and before calculating the elementary effects,
cf. \code{\link[sensitivity]{morris}}. Defaults to \code{TRUE}, which is
highly recommended if the factors have different orders of magnitude, see
\code{\link[sensitivity]{morris}}.}

\item{ode_method}{[\code{character(1)}]\cr
method to be used for solving the differential equations, see 
\code{\link[deSolve]{ode}}. Defaults to \code{"lsoda"}.}

\item{ode_parallel}{[\code{logical(1)}]\cr
logical indicating if a parallelization shall be done for computing the
\code{\link[deSolve]{ode}}-results for the different parameter combinations
generated for Monte Carlo estimation of the sensitivity indices.}

\item{ode_parallel_ncores}{[\code{integer(1)}]\cr
number of processor cores to be used for parallelization. Only applies if
\code{ode_parallel = TRUE}. If set to \code{NA} (as per default) and 
\code{ode_parallel = TRUE}, 1 processor core is used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
List of class \code{morrisRes} of length \code{length(state_init)} 
  containing in each element a matrix for one state variable. The
  matrices itself contain the Morris' SA results for all timepoints (rows: 
  \code{mu, mu.star} and \code{sigma} for every parameter; columns: 
  timepoints).
}
\description{
\code{ODEmorris.default} is the default method of \code{\link{ODEmorris}}. It
performs a sensitivity analysis for general ODE models using Morris' 
elementary effects screening method.
}
\details{
The sensitivity analysis is done for all state variables and all
  timepoints simultaneously using \code{\link[sensitivity]{morris}} from 
  the package \code{sensitivity}. \code{\link[sensitivity]{morris}} can
  handle three-dimensional arrays as output for its model function. Each 
  element of the third dimension of the output array is used to contain the 
  results for one state variable of the ODE system. Each element of the 
  second dimension of the output array is used for one timepoint. The use of
  an array as output saves time (compared to looping over all state 
  variables and all timepoints) since \code{\link[deSolve]{ode}} from the 
  package \code{deSolve} does its calculations for all state variables and 
  all timepoints anyway, so \code{\link[deSolve]{ode}} only needs to be 
  executed once.
}
\note{
\code{\link[deSolve]{ode}} sometimes can't solve an ODE system if 
  unrealistic parameter combinations are sampled by 
  \code{\link[sensitivity]{morris}}. Hence, \code{NA}s might occur in the 
  Morris sensitivity results, such that \code{\link{ODEmorris.default}} fails 
  for one or many points of time! For this reason, if \code{NA}s occur, 
  please make use of \code{\link{ODEsobol.default}} instead or
  restrict the input parameter value intervals usefully using
  \code{binf}, \code{bsup} and \code{scale = TRUE}. It might also be
  helpful to try another ODE-solver (argument \code{ode_method}). Problems 
  are known for the \code{ode_method}s \code{"euler"}, \code{"rk4"} and 
  \code{"ode45"}. In contrast, the \code{ode_method}s \code{"vode"}, 
  \code{"bdf"}, \code{"bdf_d"}, \code{"adams"}, \code{"impAdams"} and 
  \code{"impAdams_d"} might be even faster than the default 
  \code{ode_method} \code{"lsoda"}.
  
  If \code{\link[sensitivity]{morris}} throws a warning message saying
  "In ... keeping ... repetitions out of ...", try using a bigger number of 
  \code{levels} in the \code{design} argument.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007) #####
# Definition of the model itself, parameters, initial state values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {
    
    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)
    
    return(list(c(dVoltage, dCurrent)))
  })
}
FHNstate  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 50, by = 5)

FHNres <- ODEmorris(mod = FHNmod,
                    pars = c("a", "b", "s"),
                    state_init = FHNstate,
                    times = FHNtimes,
                    seed = 2015,
                    binf = c(0.18, 0.18, 2.8),
                    bsup = c(0.22, 0.22, 3.2),
                    r = 50,
                    design = list(type = "oat", levels = 100, grid.jump = 1),
                    scale = TRUE,
                    ode_method = "adams",
                    ode_parallel = TRUE,
                    ode_parallel_ncores = 2)

}
\author{
Frank Weber
}
\references{
J. O. Ramsay, G. Hooker, D. Campbell and J. Cao, 2007,
  \emph{Parameter estimation for differential equations: a generalized 
  smoothing approach}, Journal of the Royal Statistical Society, Series B, 
  69, Part 5, 741--796.
}
\seealso{
\code{\link[sensitivity]{morris}, \link{plot.morrisRes}}
}

