% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODEmorris.R
\name{ODEmorris}
\alias{ODEmorris}
\title{Morris SA for ODEs}
\usage{
ODEmorris(mod, pars, yini, times, seed = 2015, r, design = list(type =
  "oat", levels = 100, grid.jump = 1), trafo = function(Y) rowSums(Y^2),
  ncores = 1)
}
\arguments{
\item{mod}{[\code{function(Time, State, Pars)}]\cr
model to examine, cf. example below.}

\item{pars}{[\code{numeric(k)}]\cr
vector of \code{k} input variable names.}

\item{yini}{[\code{numeric(z)}]\cr
vector of \code{z} initial values.}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed
(vector of arbitrary length). Also the
first point of time must be positive.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{r}{[\code{integer(1)}]\cr
number of repetitions of the \code{design},
cf. \code{\link[sensitivity]{morris}}.}

\item{design}{[\code{list}]\cr
a list specifying the design type and its parameters,
cf. \code{\link[sensitivity]{morris}}.}

\item{trafo}{[\code{function}]\cr
function to transform \code{z > 1} output variables to
IR [only needed, if \code{z > 1}]. Must be able to deal with a
matrix.}

\item{ncores}{[\code{integer(1)}]\cr
  number of processor cores to be used for calculating the sensitivity
  indices. Must be between 1 and 4.}
}
\value{
matrix of Morris SA results (i.e. \code{mu, mu.star} and
  \code{sigma}) for every point of
  time of the \code{times} vector, of class \code{sobolRes}.
}
\description{
\code{ODEmorris} performs a sensitivity analysis for
ordinary differential equations using
Morris's elementary effects screening method.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al, 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # paramter a
              b = 0.3,     # paramter b
              s = 3)       # paramter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 100, by = 10)

FHNres <- ODEmorris(mod = FHNmod,
                    pars = names(FHNpars),
                    yini = FHNyini,
                    times = FHNtimes,
                    seed = 2015,
                    r = 25,
                    design =
                        list(type = "oat", levels = 100, grid.jump = 1),
                    trafo = function(Y) Y[, 1],    # voltage only,
                    ncores = 4)
}
\seealso{
\code{\link[sensitivity]{morris}}
}

