% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEmorris.R
\name{ODEmorris}
\alias{ODEmorris}
\title{Morris SA for ODEs}
\usage{
ODEmorris(mod, pars, yini, times, seed = 2015, binf = 0, bsup = 1,
  r = 25, design = list(type = "oat", levels = 100, grid.jump = 1),
  scale = TRUE, trafo = function(Y) rowSums(Y^2), ncores = 1)
}
\arguments{
\item{mod}{[\code{function(Time, State, Pars)}]\cr
model to examine, cf. example below.}

\item{pars}{[\code{character(k)}]\cr
vector of \code{k} input variable names.}

\item{yini}{[\code{numeric(z)}]\cr
vector of \code{z} initial values.}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed
(vector of arbitrary length). Also the
first point of time must be positive.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{binf}{[\code{numeric(k)}]\cr
vector of lower borders of possible input parameter values.
If they are all equal, a single value can be set.}

\item{bsup}{[\code{numeric(k)}]\cr
vector of upper borders of possible input parameter values.
If they are all equal, a single value can be set.}

\item{r}{[\code{integer(1)}]\cr
number of repetitions of the \code{design},
cf. \code{\link[sensitivity]{morris}}.}

\item{design}{[\code{list}]\cr
a list specifying the design type and its parameters,
cf. \code{\link[sensitivity]{morris}}.}

\item{scale}{[\code{logical(1)}]\cr
if \code{TRUE}, scaling is done for the input design of experiments after 
building the design and before calculating the elementary effects,
cf. \code{\link[sensitivity]{morris}}. Defaults to \code{FALSE}, but it is
highly recommended to use \code{scale = TRUE} if the factors have different
orders of magnitude.}

\item{trafo}{[\code{function}]\cr
function to transform \code{z > 1} output variables to
IR [only needed, if \code{z > 1}]. Must be able to deal with a
matrix.}

\item{ncores}{[\code{integer(1)}]\cr
number of processor cores to be used for calculating the sensitivity
indices. Must be between 1 and 4.}
}
\value{
list of class \code{morrisRes} with
  \itemize{
    \item \code{res}, the matrix of Morris SA results
      (i.e. \code{mu, mu.star} and \code{sigma}) for every point of
      time in the \code{times} vector and
    \item \code{pars}, the parameter names.
  }
}
\description{
\code{ODEmorris} performs a sensitivity analysis for
ordinary differential equations using
Morris's elementary effects screening method.
}
\note{
\code{\link[deSolve]{ode}} or rather its standard solver \code{lsoda}
  sometimes cannot solve an ODE system if unrealistic parameters
  are sampled by \code{\link[sensitivity]{morris}}. Hence
  \code{NA}s might occur in the Morris sensitivity results, such
  that \code{\link{ODEmorris}} fails for one or many points of time!
  For this reason, if \code{NA}s occur, please make use of
  \code{\link{ODEsobol}} instead or
  restrict the input parameter value intervals usefully using
  \code{binf} and \code{bsup}!
  
  If \code{\link[sensitivity]{morris}} throws a warning message saying
  "In ... keeping ... repetitions out of ...", try using a bigger number of 
  \code{levels} in the \code{design} argument.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # parameter a
              b = 0.3,     # parameter b
              s = 3)       # parameter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 100, by = 10)

FHNres <- ODEmorris(mod = FHNmod,
                    pars = names(FHNpars),
                    yini = FHNyini,
                    times = FHNtimes,
                    seed = 2015,
                    binf = c(0.18, 0.18, 2.8),
                    bsup = c(0.22, 0.22, 3.2),
                    r = 25,
                    design =
                        list(type = "oat", levels = 100, grid.jump = 1),
                    scale = TRUE,
                    trafo = function(Y) Y[, 1],    # voltage only
                    ncores = 4)

}
\references{
J. O. Ramsay, G. Hooker, D. Campbell and J. Cao, 2007,
  \emph{Parameter estimation for differential equations: a generalized 
  smoothing approach}, Journal of the Royal Statistical Society, Series B, 
  69, Part 5, 741--796.
}
\seealso{
\code{\link[sensitivity]{morris}},
  \code{\link{plot.morrisRes}}
}

