% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODEsensitivity-package.R
\docType{package}
\name{ODEsensitivity}
\alias{ODEsensitivity}
\alias{ODEsensitivity-package}
\title{Performing Sensitivity Analysis in ODE-Models}
\description{
ODEsensitivity provides some useful functions to analyse and
visualisize the results of Sensitivity Analyses in ODE-models.
}
\details{
This package is an addition to the
\link{sensitivity}-package.
}

