% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEsobol.default.R
\name{ODEsobol.default}
\alias{ODEsobol.default}
\title{Sobol' SA for General ODE Models}
\usage{
\method{ODEsobol}{default}(mod, pars, state_init, times, seed = 2015,
  n = 1000, rfuncs = "runif", rargs = "min = 0, max = 1",
  sobol_method = "martinez", ode_method = "lsoda", ode_parallel = FALSE,
  ode_parallel_ncores = NA, ...)
}
\arguments{
\item{mod}{[\code{function(Time, State, Pars)}]\cr
model to examine, cf. example below.}

\item{pars}{[\code{character(k)}]\cr
vector of \code{k} input variable names.}

\item{state_init}{[\code{numeric(z)}]\cr
vector of \code{z} initial values. Must be named (with unique names).}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed (vector of arbitrary 
length). The first point of time must be greater than zero.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{n}{[\code{integer(1)}]\cr
number of random parameter values (\code{n} per input factor) used to 
estimate the variance-based sensitivity indices by Monte Carlo method.
(Variance-based methods for sensitivity analysis rely on 
Monte-Carlo-simulation to estimate the integrals needed for the calculation
of the sensitivity indices.) Defaults to 1000.}

\item{rfuncs}{[\code{character(1} or \code{k)}]\cr
names of the functions used to generate the \code{n} random values
for the \code{k} parameters. Can be of length 1 or \code{k}. If of length 
1, the same function is used for all parameters. Defaults to 
\code{"runif"}, so a uniform distribution is assumed for all parameters.}

\item{rargs}{[\code{character(1} or \code{k)}]\cr
arguments to be passed to the functions in \code{rfuncs}. Can be of length 
1 or \code{k}. If of length 1, the same arguments are used for all 
parameters. Each element of \code{rargs} has to be a string of the form 
\code{"tag1 = value1, tag2 = value2, ..."}, see example below. Default is 
\code{"min = 0, max = 1"}, so (together with the default value of 
\code{rfuncs}) a uniform distribution on [0, 1] is assumed for all 
parameters.}

\item{sobol_method}{[\code{character(1)}]\cr
either \code{"jansen"} or \code{"martinez"}, specifying which modification
of the variance-based Sobol' method shall be used. Defaults to 
\code{"martinez"}.}

\item{ode_method}{[\code{character(1)}]\cr
method to be used for solving the differential equations, see 
\code{\link[deSolve]{ode}}. Defaults to \code{"lsoda"}.}

\item{ode_parallel}{[\code{logical(1)}]\cr
logical indicating if a parallelization shall be done for computing the
\code{\link[deSolve]{ode}}-results for the different parameter combinations
generated for Monte Carlo estimation of the SA indices.}

\item{ode_parallel_ncores}{[\code{integer(1)}]\cr
number of processor cores to be used for parallelization. Only applies if
\code{ode_parallel = TRUE}. If set to \code{NA} (as per default) and 
\code{ode_parallel = TRUE}, 1 processor core is used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
List of length \code{length(state_init)} and of class \code{sobolRes} 
  containing in each element a list of the Sobol' SA results for the 
  corresponding \code{state_init}-variable (i.e. first order sensitivity 
  indices \code{S} and total sensitivity indices \code{T}) for every point of
  time of the \code{times} vector.
}
\description{
\code{ODEsobol.default} is the default method of \code{\link{ODEmorris}}. It
performs a variance-based sensitivity analysis for general ODE models.
}
\details{
The sensitivity analysis is done for all state variables and all
  timepoints simultaneously using either
  \code{\link[sensitivity]{soboljansen}} or
  \code{\link[sensitivity]{sobolmartinez}} from the package 
  \code{sensitivity} (depending on \code{sobol_method}). 
  \code{\link[sensitivity]{soboljansen}} and
  \code{\link[sensitivity]{sobolmartinez}} can handle three-dimensional
  arrays as output for their model functions. Each element of the third 
  dimension of the output array is used to contain the results for one 
  state variable of the ODE system. Each element of the second dimension of
  the output array is used for one timepoint. The use of an array as output
  saves time (compared to looping over all state variables and all 
  timepoints) since \code{\link[deSolve]{ode}} from the package 
  \code{deSolve} does its calculations for all state variables and all
  timepoints anyway, so \code{\link[deSolve]{ode}} only needs to be executed 
  once.
}
\note{
It might be helpful to try different types of ODE-solvers (argument 
  \code{ode_method}). Problems are known for the \code{ode_method}s 
  \code{"euler"}, \code{"rk4"} and \code{"ode45"}. 
  In contrast, the \code{ode_method}s \code{"vode"}, \code{"bdf"}, 
  \code{"bdf_d"}, \code{"adams"}, \code{"impAdams"} and \code{"impAdams_d"} 
  might be even faster than the standard \code{ode_method} \code{"lsoda"}.
  
  If \code{n} is too low, the Monte Carlo estimation of the sensitivity 
  indices might be very bad and even produce first order indices < 0 or
  total indices > 1. First order indices in the interval [-0.05, 0) and total 
  indices in (1, 1.05] are considered as minor deviations and set to 0 
  resp. 1 without a warning. First order indices < -0.05 or total indices 
  > 1.05 are considered as major deviations. They remain unchanged and a 
  warning is thrown. Up to now, first order indices > 1 or total indices < 0
  haven't occured yet. If this should be the case, please contact the package
  author.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007) #####
# Definition of the model itself, parameters, initial state values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {
    
    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)
    
    return(list(c(dVoltage, dCurrent)))
  })
}
FHNstate  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 50, by = 5)

# Warning: The following code might take a long time!

# Arguments "rfuncs" and "rargs" being of length 1:
FHNres_default <- ODEsobol(mod = FHNmod,
                           pars = c("a", "b", "s"),
                           state_init = FHNstate,
                           times = FHNtimes,
                           seed = 2015,
                           n = 1000,
                           rfuncs = "runif",
                           rargs = "min = 0.1, max = 1.5",
                           sobol_method = "martinez",
                           ode_method = "adams",
                           ode_parallel = TRUE,
                           ode_parallel_ncores = 2)

# Arguments "rfuncs" and "rargs" being of the same length as "pars" (the
# distributions and their corresponding arguments are chosen more or less 
# arbitrarily and shall only demonstrate the use of different distributions):
FHNres <- ODEsobol(mod = FHNmod,
                   pars = c("a", "b", "s"),
                   state_init = FHNstate,
                   times = FHNtimes,
                   seed = 2015,
                   n = 1000,
                   rfuncs = c("runif", "rnorm", "rexp"),
                   rargs = c("min = 0.18, max = 0.22", 
                             "mean = 0.2, sd = 0.2 / 3",
                             "rate = 1 / 3"),
                   sobol_method = "martinez",
                   ode_method = "adams",
                   ode_parallel = TRUE,
                   ode_parallel_ncores = 2)

}
\author{
Frank Weber
}
\references{
J. O. Ramsay, G. Hooker, D. Campbell and J. Cao, 2007,
  \emph{Parameter estimation for differential equations: a generalized 
  smoothing approach}, Journal of the Royal Statistical Society, Series B, 
  69, Part 5, 741--796.
}
\seealso{
\code{\link[sensitivity]{soboljansen},
  \link[sensitivity]{sobolmartinez},
  \link{plot.sobolRes}}
}

