% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEsobol.R
\name{ODEsobol}
\alias{ODEsobol}
\title{Sobol SA for ODEs}
\usage{
ODEsobol(mod, pars, yini, times, seed = 2015, n = 1000,
  trafo = function(Y) rowSums(Y^2), ncores = 1)
}
\arguments{
\item{mod}{[\code{function(Time, State, Pars)}]\cr
model to examine, cf. example below.}

\item{pars}{[\code{character(k)}]\cr
vector of \code{k} input variable names.}

\item{yini}{[\code{numeric(z)}]\cr
vector of \code{z} initial values.}

\item{times}{[\code{numeric}]\cr
points of time at which the SA should be executed
(vector of arbitrary length). Also the
first point of time must be positive.}

\item{seed}{[\code{numeric(1)}]\cr
seed.}

\item{n}{[\code{integer(1)}]\cr
parameter \code{nboot} used in \code{\link{soboljansen}},
i.e. the number of bootstrap replicates.}

\item{trafo}{[\code{function}]\cr
function to transform \code{z > 1} output variables to
IR [only needed, if \code{z > 1}]. Must be able to deal with a
matrix.}

\item{ncores}{[\code{integer(1)}]\cr
number of processor cores to be used for calculating the sensitivity
indices. Must be between 1 and 4.}
}
\value{
list of Sobol SA results (i.e. 1st order sensitivity indices
  \code{S} and total sensitivity indices \code{T}) for every point of
  time of the \code{times} vector, of class \code{sobolRes}.
}
\description{
\code{ODEsobol} performs a sensitivity analysis for
ordinary differential equations using
the variance-based Sobol method.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al, 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # parameter a
              b = 0.3,     # parameter b
              s = 3)       # parameter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 20, by = 0.5)

FHNres <- ODEsobol(mod = FHNmod,
                   pars = names(FHNpars),
                   yini = FHNyini,
                   times = FHNtimes,
                   seed = 2015,
                   n = 10,                        # use n >> 10!
                   trafo = function(Y) Y[, 1],    # voltage only
                   ncores = 4)

}
\seealso{
\code{\link[sensitivity]{sobol}},
  \code{\link[sensitivity]{soboljansen}},
  \code{\link{plot.sobolRes}}
}

