% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODEsobol.R
\name{ODEsobol}
\alias{ODEsobol}
\title{Sobol SA for ODEs}
\usage{
ODEsobol(mod, pars, yini, times, seed = 2015, n = 1000,
  trafo = function(Y) rowSums(Y^2))
}
\arguments{
\item{mod}{model to examine.}

\item{pars}{vector of input variable names.}

\item{yini}{vector of initial values.}

\item{times}{points of time at which the SA should be executed.}

\item{seed}{seed.}

\item{n}{parameter \code{nboot} used in \code{\link{soboljansen}},
i.e. the number of bootstrap replicates.}

\item{trafo}{function to transform \code{z > 1} output variables to
  IR [only needed, if \code{z > 1}]. Must be able to deal with a
  matrix.}
}
\value{
list of Sobol SA results (i.e. 1st order sensitivity indices
  \code{S} and total sensitivity indices \code{T}) for every point of
  time of the \code{times} vector, of class \code{sobolRes}.
}
\description{
\code{ODEsobol} performs a sensitivity analysis for
ordinary differential equations using
the variance-based Sobol method.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al, 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # paramter a
              b = 0.3,     # paramter b
              s = 3)       # paramter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 20, by = 0.5)

FHNres <- ODEsobol(mod = FHNmod,
                   pars = c("a", "b", "s"),
                   yini = FHNyini,
                   times = FHNtimes,
                   seed = 2015,
                   n = 10,
                   trafo = function(Y) rowSums(Y^2))
}
\seealso{
\code{\link[sensitivity]{sobol}},
  \code{\link[sensitivity]{soboljansen}}
}

