% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morrisRes.R
\name{plot.morrisRes}
\alias{plot.morrisRes}
\title{Plot of the results of Morris SA for objects of class \code{morrisRes}}
\usage{
\method{plot}{morrisRes}(x, state_plot = names(x)[1], type = "sep",
  colors_pars = NULL, main_title = NULL, legendPos = "outside", ...)
}
\arguments{
\item{x}{[\code{morrisRes}]\cr
resulting output of \code{\link{ODEmorris}}, of class \code{morrisRes}.}

\item{state_plot}{[\code{character(1)}]\cr
name of the state variable to be plotted. Defaults to the name of the
first state variable.}

\item{type}{[\code{character(1)}]\cr
plot type, choose between \code{"sep"} and \code{"trajec"}.}

\item{colors_pars}{[\code{character(>= k)}]\cr
vector of the colors to be used for the \code{k} different parameters 
(where \code{k} is the length of the vector returned by 
\code{\link[ODEnetwork]{createParamVec}} applied to the \code{ODEnetwork}-
object for which the sensitivity analysis was done). Must be at least of
length \code{k}. If \code{NULL} (the default), \code{rainbow(k)} is used.}

\item{main_title}{[\code{character(1)}]\cr
title for the plot. If \code{type = "sep"}, this is the overall title for
the two separate plots. Defaults to NULL, so a standard title is generated.}

\item{legendPos}{[\code{character(1)}]\cr
keyword for the legend position, either one of those specified in
\code{\link{legend}} or "outside", which means the legend is placed under
the plot (useful, if there are many parameters are in the model). Default 
is \code{"outside"}.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
\code{TRUE} (invisible; for testing purposes).
}
\description{
\code{plot.morrisRes} plots the results of Morris SA for objects of class 
\code{morrisRes}.
}
\details{
\code{plot} with \code{type = "sep"} plots mu.star and
  sigma separately versus time.

\code{plot} with \code{type = "trajec"} plots mu.star versus
  sigma for every point of time.
}
\note{
Unfortunately, the passing of arguments (e.g. "main") does not work
  correctly.
}
\examples{
library(ODEnetwork)
masses <- c(1, 1)
dampers <- diag(c(1, 1))
springs <- diag(c(1, 1))
springs[1, 2] <- 1
distances <- diag(c(0, 2))
distances[1, 2] <- 1
odenet <- ODEnetwork(masses, dampers, springs, 
                     cartesian = TRUE, distances = distances)
odenet <- setState(odenet, c(0.5, 1), c(0, 0))

ODEtimes <- seq(0.01, 20, by = 0.1)
ODEbinf <- c(rep(0.001, 9), -4, 0.001, -10)
ODEbsup <- c(2, 1.5, 6, 4, 6, 10, 2, 1.5, 6, -0.001, 6, -0.001)

ODEres <- ODEmorris(odenet, ODEtimes, ode_method = "adams", seed = 2015, 
                    binf = ODEbinf, bsup = ODEbsup, r = 20)

# Standard (separate plots):
plot(ODEres, state_plot = "x.2", type = "sep", legendPos = "topleft")
plot(ODEres, state_plot = "x.2", type = "sep", legendPos = "outside")
# Palette "Dark2" from the package "RColorBrewer" with some 
# additional colors:
my_cols <- c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E", 
             "#E6AB02", "#A6761D", "#666666", "black", "firebrick",
             "darkblue", "darkgreen")
plot(ODEres, state_plot = "x.2", type = "sep", colors_pars = my_cols, 
     legendPos = "outside")

# Trajectories:
plot(ODEres, state_plot = "x.2", type = "trajec", legendPos = "topleft")
plot(ODEres, state_plot = "x.2", type = "trajec", legendPos = "outside")
plot(ODEres, state_plot = "x.2", type = "trajec", colors_pars = my_cols, 
     legendPos = "outside")

}
\author{
Frank Weber
}
\seealso{
\code{\link{ODEmorris}},
  \code{\link[sensitivity]{morris_list}}
}

