% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.morrisRes.R
\name{plot.morrisRes}
\alias{plot.morrisRes}
\title{Plotting the Results of Morris SA}
\usage{
\method{plot}{morrisRes}(res, type = "sep", ...)
}
\arguments{
\item{res}{[\code{morrisRes}]\cr
resulting output of \code{\link{ODEmorris}}, of class \code{morrisRes}.}

\item{type}{[\code{character(1)}]\cr
plot type, choose between \code{"sep"} and \code{"trajec"}.}

\item{...}{additional arguments.}
}
\description{
\code{plot} plots the results of Morris SA.
}
\details{
\code{plot} with \code{type = "sep"} plots mu.star and
  sigma seperately versus time.

\code{plot} with \code{type = "trajec"} plots mu.star versus
  sigma for every point of time.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al, 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # paramter a
              b = 0.3,     # paramter b
              s = 3)       # paramter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 100, by = 10)

FHNres <- ODEmorris(mod = FHNmod,
                    pars = c("a", "b", "s"),
                    yini = FHNyini,
                    times = FHNtimes,
                    seed = 2015,
                    r = 25,
                    design =
                        list(type = "oat", levels = 100, grid.jump = 1),
                    trafo = function(Y) rowSums(Y^2))

# Plots:
plot(FHNres, type = "sep")
plot(FHNres, type = "trajec")
}
\author{
Stefan Theers
}
\seealso{
\code{\link[sensitivity]{morris}}
}

