% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ODEmorris.R
\name{plot.ODEmorris}
\alias{plot.ODEmorris}
\title{Plot of the Results of Morris Screening for Objects of Class \code{ODEmorris}}
\usage{
\method{plot}{ODEmorris}(x, pars_plot = NULL, state_plot = names(x)[1],
  kind = "sep", colors_pars = NULL, main_title = NULL,
  legendPos = "outside", type = "l", ...)
}
\arguments{
\item{x}{[\code{ODEmorris}]\cr
output of \code{\link{ODEmorris}} (of class \code{ODEmorris}).}

\item{pars_plot}{[\code{character(k)}]\cr
names of the \code{k} parameters to be plotted. If \code{NULL} (the 
default), all parameters are plotted.}

\item{state_plot}{[\code{character(1)}]\cr
name of the state variable to be plotted. Defaults to the name of the
first state variable.}

\item{kind}{[\code{character(1)}]\cr
kind of the plot, choose between \code{"sep"} and \code{"trajec"}.}

\item{colors_pars}{[\code{character(>= k)}]\cr
vector of the colors to be used for the \code{k} different parameters. Must
be at least of length \code{k} (only the first \code{k} elements will be
used, though). If \code{NULL} (the default), \code{rainbow(k)} is used.}

\item{main_title}{[\code{character(1)}]\cr
title for the plot. If \code{kind = "sep"}, this is the overall title for
the two separate plots. If \code{NULL} (the default), a standard title is 
generated.}

\item{legendPos}{[\code{character(1)}]\cr
keyword for the legend position, either one of those specified in
\code{\link{legend}} or \code{"outside"} (the default), which means the 
legend is placed under the plot (useful, if there are many parameters in 
the model).}

\item{type}{[\code{character(1)}]\cr
plot type, i.e. \code{"p", "l", "b", "c", "o", "s", "h"} or \code{"n"}. 
Defaults to \code{"l"}.}

\item{...}{additional arguments passed to \code{\link{plot.default}}.}
}
\value{
\code{TRUE} (invisible; for testing purposes).
}
\description{
\code{plot.ODEmorris} plots the results of Morris screening for objects of 
  class \code{ODEmorris}.
}
\details{
\code{plot.ODEmorris} with \code{kind = "sep"} plots mu.star and
  sigma separately versus time.

  \code{plot.ODEmorris} with \code{kind = "trajec"} plots mu.star versus
  sigma for every point of time.
}
\note{
Not all plotting arguments can be passed by \code{...}, for example
  \code{xlab} and \code{ylab} are fixed.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007) #####
# Definition of the model itself, parameters, initial state values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {
    
    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)
    
    return(list(c(dVoltage, dCurrent)))
  })
}
FHNstate  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 50, by = 5)

set.seed(4628)
FHNres <- ODEmorris(mod = FHNmod,
                    pars = c("a", "b", "s"),
                    state_init = FHNstate,
                    times = FHNtimes,
                    binf = c(0.18, 0.18, 2.8),
                    bsup = c(0.22, 0.22, 3.2),
                    r = 50,
                    design = list(type = "oat", levels = 100, grid.jump = 1),
                    scale = TRUE,
                    ode_method = "adams",
                    parallel_eval = TRUE,
                    parallel_eval_ncores = 2)

# Define custom colors for the plot:
my_cols <- c("firebrick", "chartreuse3", "dodgerblue")
plot(FHNres, state_plot = "Current", kind = "sep", colors_pars = my_cols)
plot(FHNres, state_plot = "Current", kind = "trajec", colors_pars = my_cols)

##### A network of ordinary differential equations #####
# Definition of the network using the package "ODEnetwork":
library(ODEnetwork)
masses <- c(1, 1)
dampers <- diag(c(1, 1))
springs <- diag(c(1, 1))
springs[1, 2] <- 1
distances <- diag(c(0, 2))
distances[1, 2] <- 1
lfonet <- ODEnetwork(masses, dampers, springs, 
                     cartesian = TRUE, distances = distances)
lfonet <- setState(lfonet, c(0.5, 1), c(0, 0))
LFOpars <- c("k.1", "k.2", "k.1.2")
LFOtimes <- seq(0.01, 20, by = 0.1)
LFObinf <- rep(0.001, 3)
LFObsup <- c(6, 6, 3)

set.seed(4628)
LFOres <- ODEmorris(lfonet,
                    LFOpars,
                    LFOtimes,
                    binf = LFObinf,
                    bsup = LFObsup,
                    r = 50,
                    design = list(type = "oat", levels = 100, grid.jump = 1),
                    scale = TRUE,
                    ode_method = "adams",
                    parallel_eval = TRUE,
                    parallel_eval_ncores = 2)

plot(LFOres, state_plot = "x.2", kind = "sep", colors_pars = my_cols)
plot(LFOres, state_plot = "x.2", kind = "trajec", colors_pars = my_cols)

}
\author{
Stefan Theers, Frank Weber
}
\seealso{
\code{\link{ODEmorris}, \link[sensitivity]{morris}}
}

