% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.sobolRes.R
\name{plot.sobolRes}
\alias{plot.sobolRes}
\title{Plotting the results of Sobol SA}
\usage{
\method{plot}{sobolRes}(res = LVres, type = "b", legendPos = "topleft",
  ...)
}
\arguments{
\item{res}{resulting output of \code{\link{ODEsobol}}.}

\item{type}{plot type, i.e. \code{"p", "l", "b", "c"} or \code{"n"}.}

\item{legendPos}{legend position, default is \code{"topleft"}.}

\item{...}{additional arguments.}
}
\description{
\code{plot.sobolRes} plots the results of Sobol SA.
}
\details{
1st order and total Sobol SA indices are plotted for each input
parameter against time.
}
\examples{
# none.
}
\author{
Stefan Theers
}
\seealso{
\code{\link[sensitivity]{sobol}},
  \code{\link[sensitivity]{sobol2007}}
}

