% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sobolRes.R
\name{plot.sobolRes}
\alias{plot.sobolRes}
\title{Plotting the results of Sobol SA}
\usage{
\method{plot}{sobolRes}(x, type = "b", legendPos = "topleft",
  overall_main = NULL, ...)
}
\arguments{
\item{x}{[\code{sobolRes}]\cr
resulting output of \code{\link{ODEsobol}}, of class \code{sobolRes}.}

\item{type}{[\code{character(1)}]\cr
plot type, i.e. \code{"p", "l", "b", "c", "o", "s", "h"} or \code{"n"}. 
Defaults to \code{"b"}.}

\item{legendPos}{[\code{character(1)}]\cr
legend position, default is \code{"topleft"}.}

\item{overall_main}{[\code{character(1)}]\cr
the common title for the two graphics. Default is \code{NULL}, which means
an automatic title is generated.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
TRUE (invisible; for testing purposes).
}
\description{
\code{plot.sobolRes} plots the results of Sobol SA for objects of class 
\code{sobolRes}.
}
\details{
1st order and total Sobol SA indices are plotted for each input
parameter against time.
}
\note{
Unfortunately, the passing of arguments (e.g. "main") does not work
  correctly.
}
\examples{

##### FitzHugh-Nagumo equations (Ramsay et al., 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 50, by = 5)

FHNres <- ODEsobol(mod = FHNmod,
                   pars = c("a", "b", "s"),
                   yini = FHNyini,
                   times = FHNtimes,
                   seed = 2015,
                   n = 10,                        # use n >> 10!
                   rfuncs = c("runif", "runif", "rnorm"),
                   rargs = c(rep("min = 0.18, max = 0.22", 2),
                             "mean = 3, sd = 0.2 / 3"),
                   method = "martinez",
                   nboot = 0,
                   trafo = function(Y) Y[, 1],    # voltage only
                   ncores = 2)

# Plot:
plot(FHNres, type = "l", legendPos = "topright")

}
\author{
Stefan Theers
}
\seealso{
\code{\link{ODEsobol}},
  \code{\link[sensitivity]{sobol}},
  \code{\link[sensitivity]{sobol2007}}
}

