% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.sobolRes.R
\name{plot.sobolRes}
\alias{plot.sobolRes}
\title{Plotting the results of Sobol SA}
\usage{
\method{plot}{sobolRes}(res, type = "b", legendPos = "topleft", ...)
}
\arguments{
\item{res}{[\code{sobolRes}]\cr
resulting output of \code{\link{ODEsobol}}, of class \code{sobolRes}.}

\item{type}{[\code{character(1)}]\cr
plot type, i.e. \code{"p", "l", "b", "c"} or \code{"n"}.}

\item{legendPos}{[\code{character(1)}]\cr
legend position, default is \code{"topleft"}.}

\item{...}{additional arguments.}
}
\description{
\code{plot.sobolRes} plots the results of Sobol SA.
}
\details{
1st order and total Sobol SA indices are plotted for each input
parameter against time.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al, 2007)
# definition of the model itself, parameters, initial values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {

    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)

    return(list(c(dVoltage, dCurrent)))
  })
}

FHNpars  <- c(a = 0.2,     # paramter a
              b = 0.3,     # paramter b
              s = 3)       # paramter s (= c in the original notation)

FHNyini  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 20, by = 0.5)

FHNres <- ODEsobol(mod = FHNmod,
                   pars = names(FHNpars),
                   yini = FHNyini,
                   times = FHNtimes,
                   seed = 2015,
                   n = 10,                        # use n >> 10!
                   trafo = function(Y) Y[, 1],    # voltage only
                   ncores = 4)

# Plot:
plot(FHNres, type = "l", legendPos = "topright")
}
\author{
Stefan Theers
}
\seealso{
\code{\link{ODEsobol}},
  \code{\link[sensitivity]{sobol}},
  \code{\link[sensitivity]{sobol2007}}
}

