% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sobolRes.R
\name{plot.sobolRes}
\alias{plot.sobolRes}
\title{Plotting the results of Sobol' SA for objects of class \code{sobolRes}}
\usage{
\method{plot}{sobolRes}(x, state_plot = names(x)[1], colors_pars = NULL,
  main_title = NULL, legendPos = "outside", type = "l", ...)
}
\arguments{
\item{x}{[\code{sobolRes}]\cr
resulting output of \code{\link{ODEsobol}}, of class \code{sobolRes}.}

\item{state_plot}{[\code{character(1)}]\cr
name of the state variable to be plotted. Defaults to the name of the
first state variable.}

\item{colors_pars}{[\code{character(>= k)}]\cr
vector of the colors to be used for the \code{k} different parameters. Must
be at least of length \code{k}. If \code{NULL} (the default), 
\code{rainbow(k)} is used.}

\item{main_title}{[\code{character(1)}]\cr
common title for the two graphics. Default is \code{NULL}, which means
an automatic title is generated.}

\item{legendPos}{[\code{character(1)}]\cr
keyword for the legend position, either one of those specified in
\code{\link{legend}} or \code{"outside"} (the default), which means the 
legend is placed under the plot (useful, if there are many parameters in 
the model).}

\item{type}{[\code{character(1)}]\cr
plot type, i.e. \code{"p", "l", "b", "c", "o", "s", "h"} or \code{"n"}. 
Defaults to \code{"l"}.}

\item{...}{additional arguments passed to \code{\link{plot.default}}.}
}
\value{
TRUE (invisible; for testing purposes).
}
\description{
\code{plot.sobolRes} plots the results of Sobol' SA for objects of class 
  \code{sobolRes}.
}
\details{
First order and total Sobol' SA indices are plotted for one state variable 
  (chosen by argument \code{state_plot}) and all of the parameters against 
  time.
}
\note{
Not all arguments of \code{\link{plot.default}} can be passed by 
  \code{...}, for example \code{xlab} and \code{ylab} are fixed.
}
\examples{
##### FitzHugh-Nagumo equations (Ramsay et al., 2007) #####
# Definition of the model itself, parameters, initial state values
# and the times vector:
FHNmod <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {
    
    dVoltage <- s * (Voltage - Voltage^3 / 3 + Current)
    dCurrent <- - 1 / s *(Voltage - a + b * Current)
    
    return(list(c(dVoltage, dCurrent)))
  })
}
FHNstate  <- c(Voltage = -1, Current = 1)
FHNtimes <- seq(0.1, 50, by = 5)

# Warning: The following code might take a long time!
FHNres <- ODEsobol(mod = FHNmod,
                   pars = c("a", "b", "s"),
                   state_init = FHNstate,
                   times = FHNtimes,
                   seed = 2015,
                   n = 1000,
                   rfuncs = c("runif", "rnorm", "rexp"),
                   rargs = c("min = 0.18, max = 0.22", 
                             "mean = 0.2, sd = 0.2 / 3",
                             "rate = 1 / 3"),
                   sobol_method = "Martinez",
                   ode_method = "adams",
                   ode_parallel = TRUE,
                   ode_parallel_ncores = 2)

# Define custom colors for the plot:
my_cols <- c("firebrick", "chartreuse3", "dodgerblue")
plot(FHNres, state_plot = "Current", colors_pars = my_cols)

##### A network of ordinary differential equations #####
# Definition of the network using the package "ODEnetwork":
library(ODEnetwork)
masses <- c(1, 1)
dampers <- diag(c(1, 1))
springs <- diag(c(1, 1))
springs[1, 2] <- 1
distances <- diag(c(0, 2))
distances[1, 2] <- 1
lfonet <- ODEnetwork(masses, dampers, springs, 
                     cartesian = TRUE, distances = distances)
lfonet <- setState(lfonet, c(0.5, 1), c(0, 0))
LFOpars <- c("k.1", "k.2", "k.1.2")
LFOtimes <- seq(0.01, 20, by = 0.1)

LFOres <- ODEsobol(lfonet, 
                   LFOpars, 
                   LFOtimes, 
                   seed = 2015, 
                   n = 1000,
                   rfuncs = c("runif", "rnorm", "rexp"),
                   rargs = c("min = 0.001, max = 6",
                             "mean = 3, sd = 0.5",
                             "rate = 1 / 3"),
                   sobol_method = "Martinez",
                   ode_method = "adams",
                   ode_parallel = TRUE,
                   ode_parallel_ncores = 2)
# (A warning is thrown, concerning the state variables "v.1" and "v.2". 
# Assuming that we are only interested in a sensitivity analysis of "x.1" and
# "x.2", this warning is ignored.)

plot(LFOres, state_plot = "x.1", colors_pars = my_cols)
plot(LFOres, state_plot = "x.2", colors_pars = my_cols)

}
\author{
Frank Weber
}
\seealso{
\code{\link{ODEsobol}, \link[sensitivity]{soboljansen},
\link[sensitivity]{sobolmartinez}}
}

